+++ src/tempname/mockTribe.ts
